+++ src/app/public/modules/list-filters/list-filter-button.component.ts
@@ -24,6 +24,8 @@ export class SkyListFilterButtonComponent implements AfterViewInit {
   })
   private filterButtonTemplate: TemplateRef<any>;
 
+  private filterButtonItemToolbarIndex: number = 5000;
+
   constructor(
     private dispatcher: ListStateDispatcher
   ) { }
@@ -33,10 +35,10 @@ export class SkyListFilterButtonComponent implements AfterViewInit {
       [
         new ListToolbarItemModel({
           template: this.filterButtonTemplate,
-          location: 'left'
+          location: 'left',
+          index: this.filterButtonItemToolbarIndex
         })
-      ],
-      1
+      ]
     );
   }
 }